feat(plugin-visual-copy): visual indicator of newly copied line
note:
previously, there's no visual distinction to indicate which new line was
adding. now, there's a visual distinction to indicate such a thing.

it's almost similar to how `gv` normal mode mapping works, the
difference is that, instead of re-select previously selected visual
selection, we select the recently copied line.

still not sure if this is a good thing or not. maybe i need to make a
few adjustment for this in the future.